StaticAnalyzer/requirements.txt
pefile==2024.8.26
yara-python==4.5.2